Fix validation in UserSubSerializer

+++ backend/api/views.py
@@ -1,5 +1,3 @@
-from itertools import chain
-
 from api.paginators import LimitPageNumberPagination
 from api.permissions import IsAdminOrReadOnly, IsAuthorOrAdminOrReadOnly
 from api.renders import IngredientDataRendererTXT
@@ -12,7 +10,7 @@ from django.shortcuts import get_object_or_404
 from djoser.views import UserViewSet
 from recipes.models import (Favorites, Ingredient, Recipe, ShoppingCart,
                             Subscriptions, Tag)
-from rest_framework import permissions, status, viewsets
+from rest_framework import permissions, status, viewsets, mixins
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from django.conf import settings
@@ -42,30 +40,42 @@ class UserViewSet(UserViewSet, viewsets.ModelViewSet):
     def subscribe(self, request, *args, **kwargs):
         pk = kwargs.get('id',)
         author = get_object_or_404(User, pk=pk)
-        user = request.user
+        user = self.request.user
         if request.method == 'POST':
-            if author == user:
-                return Response(
-                    {'errors': 'Вы не можете подписаться на себя'},
-                    status=status.HTTP_400_BAD_REQUEST)
-            if Subscriptions.objects.filter(author=author, user=user).exists():
-                return Response(
-                    {'errors': 'Вы подписаны на этого пользователя'},
-                    status=status.HTTP_400_BAD_REQUEST)
-            sub = Subscriptions(author=author, user=user)
-            sub.save()
+
+            # if author == user:
+            #     return Response(
+            #         {'errors': 'Вы не можете подписаться на себя'},
+            #         status=status.HTTP_400_BAD_REQUEST)
+            # if Subscriptions.objects.filter(author=author, user=user).exists():
+            #     return Response(
+            #         {'errors': 'Вы подписаны на этого пользователя'},
+            #         status=status.HTTP_400_BAD_REQUEST)
+            # sub = Subscriptions(author=author, user=user)
+            # sub.save()
+            # data = {
+            #     'email': author.email,
+            #     'username': author.username,
+            #     'first_name': author.first_name,
+            #     'last_name': author.last_name,
+            # }
+
             serializer = UserSubSerializer(
-                author, context={'request': request}
+                author, data=request.data, context={'request': request,
+                                            'user': user}
             )
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            if serializer.is_valid():
+                sub = Subscriptions(author=author, user=user)
+                sub.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         if request.method == 'DELETE':
-            try:
+            if Subscriptions.objects.filter(author=author, user=user).exists():
                 Subscriptions.objects.get(author=author, user=user).delete()
-            except Subscriptions.DoesNotExist:
-                return Response(
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(
                     {'errors': f'Вы не подписаны на пользователя c ID: {pk}'},
                     status=status.HTTP_400_BAD_REQUEST)
-            return Response(status=status.HTTP_204_NO_CONTENT)
         return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
     @action(methods=('GET',),
@@ -112,13 +122,12 @@ class RecipeViewSet(viewsets.ModelViewSet):
             serializer = ShortRecipeSerializer(recipe)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         if request.method == 'DELETE':
-            try:
+            if Favorites.objects.filter(user=user, recipe=recipe).exists():
                 Favorites.objects.get(user=user, recipe=recipe).delete()
-            except Favorites.DoesNotExist:
-                return Response(
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(
                     {'errors': 'Рецепта нет в избранном'},
                     status=status.HTTP_400_BAD_REQUEST)
-            return Response(status=status.HTTP_204_NO_CONTENT)
         return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
     @action(methods=('POST', 'DELETE'),
@@ -139,13 +148,12 @@ class RecipeViewSet(viewsets.ModelViewSet):
             serializer = ShortRecipeSerializer(recipe)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         if request.method == 'DELETE':
-            try:
+            if ShoppingCart.objects.filter(user=user, recipe=recipe).exists():
                 ShoppingCart.objects.get(user=user, recipe=recipe).delete()
-            except ShoppingCart.DoesNotExist:
-                return Response(
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(
                     {'errors': 'Рецепта нет в списке покупок'},
                     status=status.HTTP_400_BAD_REQUEST)
-            return Response(status=status.HTTP_204_NO_CONTENT)
         return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
     @action(
